+++ src/component/normalDetails/Header.jsx
