+++ src/component/normalDetails/Stats.jsx
