+++ src/component/normalDetails/Tabs.jsx
