+++ src/component/normalDetails/informacion/infoAbilities.jsx
