+++ src/component/normalDetails/informacion/infoDescription.jsx
