+++ src/component/normalDetails/informacion/infoSwiper.jsx
